ASoC: Add SSM4567 support
This patch adds support for the ssm4567. The ssm4567 is a mono class D audio
amplifier with an I2S interface for audio in. The control bus can either be I2C
or TDM-C. The chip additionally supports monitoring voltage and current of the
speaker as well as battery voltage which can be send back via the serial port to
the host processor or a DSP for processing.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>